requirements.txt
datasets==2.21.0
sentence-transformers==3.3.1
xgboost==2.0.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
joblib==1.4.2
tqdm==4.67.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
langgraph==0.2.39